requirements.txt
fastapi[standard]
pydantic
numpy
scikit-learn
sentence-transformers
python-dotenv
requests